move lock removal to coordinator
