+++ homeassistant/components/tedee/lock.py
@@ -65,11 +65,6 @@ class TedeeLockEntity(TedeeEntity, LockEntity):
         """Return true if lock is jammed."""
         return self._lock.is_state_jammed
 
-    @property
-    def available(self) -> bool:
-        """Return True if entity is available."""
-        return super().available and self._lock.is_connected
-
     async def async_unlock(self, **kwargs: Any) -> None:
         """Unlock the door."""
         try:
