+++ tests/components/tedee/test_lock.py
@@ -24,7 +24,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers import device_registry as dr, entity_registry as er
 
-from tests.common import async_fire_time_changed
+from tests.common import MockConfigEntry, async_fire_time_changed
 
 pytestmark = pytest.mark.usefixtures("init_integration")
 
@@ -209,17 +209,31 @@ async def test_update_failed(
     assert state.state == STATE_UNAVAILABLE
 
 
-async def test_lock_unavailable(
+async def test_cleanup_disconnected_locks(
     hass: HomeAssistant,
     mock_tedee: MagicMock,
+    device_registry: dr.DeviceRegistry,
+    mock_config_entry: MockConfigEntry,
     freezer: FrozenDateTimeFactory,
 ) -> None:
-    """Test update failed."""
+    """Ensure disconnected locks are cleaned up."""
+
+    devices = dr.async_entries_for_config_entry(
+        device_registry, mock_config_entry.entry_id
+    )
+
+    locks = [device.name for device in devices]
+    assert "Lock-1A2B" in locks
+
+    # remove a lock and reload integration
     mock_tedee.locks_dict.pop(12345)
     freezer.tick(timedelta(minutes=10))
     async_fire_time_changed(hass)
     await hass.async_block_till_done()
 
-    state = hass.states.get("lock.lock_1a2b")
-    assert state is not None
-    assert state.state == STATE_UNAVAILABLE
+    devices = dr.async_entries_for_config_entry(
+        device_registry, mock_config_entry.entry_id
+    )
+
+    locks = [device.name for device in devices]
+    assert "Lock-1A2B" not in locks
